goodread/requirements.txt
itemadapter==0.9.0
pymongo==4.7.2
Scrapy==2.11.2
kafka-python==2.0.2